requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
Brotli==1.1.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cssselect2==0.8.0
fastapi==0.120.4
fonttools==4.60.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
pyphen==0.17.2
sniffio==1.3.1
starlette==0.49.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
fastapi
uvicorn
jinja2
weasyprint
python-multipart
colorama
python-dotenv
black
flake8
mypy
python-dotenv
requests